fix: use u32 atomics in Stats to fix ARMv7 cross-build
ARMv7 (32-bit) only supports lock-free atomic operations on types no
wider than the native word size (32 bits). The Stats struct used
std.atomic.Value(u64) which caused two compilation errors on the
arm-linux-musleabihf target:

  error: expected 32-bit integer type or smaller; found 64-bit integer type

Switch all Stats fields and the socks5.Connection.connect_failures
pointer from u64 to u32. The counters are observability hints only
(monotonic, relaxed ordering) so wrapping at 2^32 is acceptable.

+++ src/server.zig
@@ -71,11 +71,13 @@ const Pool = struct {
 
 /// Atomic operational counters. Updated off the per-connection hot path with
 /// relaxed (monotonic) ordering — they are observability hints, not a barrier.
+/// u32 is used so the atomics compile on 32-bit targets such as ARMv7, which
+/// only support lock-free atomics for types no wider than the native word size.
 const Stats = struct {
-    accepted: std.atomic.Value(u64) = .init(0),
-    refused: std.atomic.Value(u64) = .init(0),
-    active: std.atomic.Value(u64) = .init(0),
-    connect_failures: std.atomic.Value(u64) = .init(0),
+    accepted: std.atomic.Value(u32) = .init(0),
+    refused: std.atomic.Value(u32) = .init(0),
+    active: std.atomic.Value(u32) = .init(0),
+    connect_failures: std.atomic.Value(u32) = .init(0),
 };
 
 pub const Server = struct {
